begin pointer handling in apphost

+++ primary/bitbang.hpp
@@ -82,7 +82,7 @@ static inline uint64_t bitsValueFromBytes(const uint8_t *bytes, size_t startbit,
     uint64_t value = 0;
 
     if (bigendian) {
-	    for (size_t i=bitcount; i>= 0; i--) {
+	    for (int i=bitcount; i>= 0; i--) {
 		    size_t byteoffset=0;
             size_t bitoffset=0;
             getbitbyteoffset(startbit+i, byteoffset, bitoffset);
